// src/app/generated/model/itemResponse.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ItemImageResponse } from './itemImageResponse';


export interface ItemResponse { 
    /**
     * Unique identifier of the item
     */
    id?: number;
    /**
     * Unique stock keeping unit code for the item
     */
    sku?: string;
    /**
     * Display name of the item
     */
    name?: string;
    /**
     * Category the item belongs to
     */
    category?: string;
    /**
     * Unit of measure used for stock counts
     */
    unitOfMeasure?: string;
    /**
     * Price at which the item is sold
     */
    sellingPrice?: number;
    /**
     * Optional reference-only default cost price
     */
    defaultCostPrice?: number;
    /**
     * Whether the item is currently active
     */
    active?: boolean;
    /**
     * List of images associated with the item, ordered by sortOrder
     */
    images?: Array<ItemImageResponse>;
    /**
     * Timestamp when the item was created
     */
    createdAt?: string;
    /**
     * Timestamp when the item was last updated
     */
    updatedAt?: string;
}