// src/app/generated/model/materialRequestResponse.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MaterialRequestLineItemResponse } from './materialRequestLineItemResponse';


export interface MaterialRequestResponse { 
    /**
     * Unique identifier of the material request
     */
    id?: number;
    /**
     * Identifier of the site (a Warehouse of type SITE) this request is for
     */
    siteWarehouseId?: number;
    /**
     * Name of the site this request is for
     */
    siteWarehouseName?: string;
    /**
     * ID of the user who made this request
     */
    requestedBy?: number;
    /**
     * Full name of the user who made this request
     */
    requestedByName?: string;
    /**
     * Date the materials are needed by, if known
     */
    dateNeeded?: string;
    /**
     * Current status of the request
     */
    status?: MaterialRequestResponse.StatusEnum;
    /**
     * Optional free-text notes about the request
     */
    notes?: string;
    /**
     * Line items requested
     */
    lines?: Array<MaterialRequestLineItemResponse>;
    /**
     * Timestamp when the request was created
     */
    createdAt?: string;
    /**
     * Timestamp when the request was last updated
     */
    updatedAt?: string;
}
export namespace MaterialRequestResponse {
    export const StatusEnum = {
        Draft: 'DRAFT',
        Submitted: 'SUBMITTED',
        PartiallyFulfilled: 'PARTIALLY_FULFILLED',
        Fulfilled: 'FULFILLED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}